requirements.txt
grpcio-tools>=1.60.0
grpcio>=1.60.0
protobuf>=4.25.0
temporalio>=1.18.1
